Use 32-bit integers for machine pointers and values
It's clear from decode() that large integers on this machine are 32-bits
and, since the same stack is used for both pointers and integers, it
doesn't make sense to treat the pointers as anything else.

I treat integers used in an address context as unsigned and those used
in an integer context as signed. There's an argument to be made for
treating them as signed in all cases and aborting when a negative
address is encountered, but this is simpler and I don't think it will
break anything.