Fix (nearly) everything (#1)
* Use 32-bit integers for machine pointers and values

It's clear from decode() that large integers on this machine are 32-bits
and, since the same stack is used for both pointers and integers, it
doesn't make sense to treat the pointers as anything else.

I treat integers used in an address context as unsigned and those used
in an integer context as signed. There's an argument to be made for
treating them as signed in all cases and aborting when a negative
address is encountered, but this is simpler and I don't think it will
break anything.

* Use std::vector::at() to check bounds for us

* Fix inverted if condition

* Don't call runtime functions from tbfcompile()

i.e. match what the Python version does.

* Replace assert(false) with equivalent abort()

* Fix incorrect math in extend_memory()

Sanity check is 100k, not 10k. Also, resize() semantics aren't the same
as Python's extend(): we need to pass the new size, not the size delta.

* Remove superfluous two's complement check

Or, at least, it would be superfluous if it weren't missing a zero in
the constant. Since we're using native integers, we can just cast to
signed instead of manually handling two's complement.

* Check length of bytes in store() like we do in literal_word()

* Remove unneccessary bit mask

Since the value's in a uint32_t, it wraps automatically.

* Properly check for EOF in read_byte()

* Use EXIT_FAILURE instead of 1 for return value

* Don't loop on EOF flag

The flag only gets set *after* a failing read, so looping on it makes us
read an extra (bogus) byte.

* Initialize run_time_dispatch globally

* Make dispatch table digit initialization more concise

* assert() that we don't pop from an empty stack

* s/end_conditional/start_conditional/

* Remove unnecessary "continue" from empty loop

* Use casts instead of bit masks to truncate ints

The compiler warns us if there's no explicit cast, and the cast makes
the bit mask unnecessary.

* Make the compiler meaner